pyproject.toml
[project]
name = "the-oracle"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.1",
    "ebooklib>=0.20",
    "fastapi>=0.133.1",
    "numpy>=2.4.2",
    "openai>=2.24.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "tiktoken>=0.12.0",
    "uvicorn>=0.41.0",
]

